Refactor: moved MuleSdkDslConstants to internal.
