+++ modules/extensions-mule-support/src/main/java/org/mule/runtime/module/extension/mule/internal/dsl/MuleSdkDslConstants.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.module.extension.mule.api.dsl;
+package org.mule.runtime.module.extension.mule.internal.dsl;
 
 import static org.mule.runtime.extension.api.util.XmlModelUtils.buildSchemaLocation;
 import static org.mule.runtime.internal.dsl.DslConstants.DEFAULT_NAMESPACE_URI_MASK;
